Move view and contentgrid/empty components to main folder.
