+++ src/components/content/ContentMain.tsx
@@ -11,9 +11,9 @@ import { selectUser } from "../../app/slices/user/userSlice";
 import { Fab } from "@rmwc/fab";
 import { DrawerAppContent } from "@rmwc/drawer";
 import { TopAppBarFixedAdjust } from "@rmwc/top-app-bar";
-import { ContentGrid } from "./ContentGrid";
-import { ContentEmpty } from "./ContentEmpty";
 import { AppBar } from "../main/app_bar/AppBar";
+import { ContentEmpty } from "../main/content/ContentEmpty";
+import { ContentGrid } from "../main/content/ContentGrid";
 import { DrawerDetails } from "../main/DrawerDetails";
 import { DrawerFilter } from "../main/DrawerFilter";
 import { DialogSales } from "../main/DialogSales";
