+++ src/components/main/content/ContentEmpty.scss
