+++ src/components/main/views/card/ElementCard.tsx
@@ -2,12 +2,12 @@ import React, { useContext } from "react";
 import Twemoji from "react-twemoji";
 import LazyLoad from "react-lazy-load";
 import classNames from "classnames";
-import { useAppSelector } from "../../../app/hooks";
-import { selectPage } from "../../../app/slices/common/commonSlice";
-import { iconObject } from "../../../app/slices/common/functions";
-import { SetType } from "../../../app/slices/main/types";
-import { selectFavorites, selectHidden, selectUser } from "../../../app/slices/user/userSlice";
-import { UserContext } from "../../../app/slices/user/contexts";
+import { useAppSelector } from "../../../../app/hooks";
+import { selectPage } from "../../../../app/slices/common/commonSlice";
+import { iconObject } from "../../../../app/slices/common/functions";
+import { SetType } from "../../../../app/slices/main/types";
+import { selectFavorites, selectHidden, selectUser } from "../../../../app/slices/user/userSlice";
+import { UserContext } from "../../../../app/slices/user/contexts";
 import { Typography } from "@rmwc/typography";
 import {
   Card,
