+++ src/components/main/views/card/ViewCard.scss
