+++ src/components/main/views/card/ViewCard.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import moment from "moment";
-import { SetType } from "../../../app/slices/main/types";
+import { SetType } from "../../../../app/slices/main/types";
 import { ElementCard } from "./ElementCard";
 import "./ViewCard.scss";
