+++ src/components/main/views/compact/ElementCompact.scss
@@ -1,7 +1,7 @@
 @use "@material/ripple/_index" as ripple;
 @use "@material/rtl/_index" as rtl;
 @use "@material/theme/_index" as theme;
-@use "../../../util";
+@use "../../../../util";
 
 .compact {
   .mdc-list.mdc-list--two-line .mdc-list-item {
