+++ src/components/main/views/compact/ElementCompact.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import Twemoji from "react-twemoji";
-import { useAppSelector } from "../../../app/hooks";
-import { selectPage } from "../../../app/slices/common/commonSlice";
-import { iconObject } from "../../../app/slices/common/functions";
-import { SetType } from "../../../app/slices/main/types";
+import { useAppSelector } from "../../../../app/hooks";
+import { selectPage } from "../../../../app/slices/common/commonSlice";
+import { iconObject } from "../../../../app/slices/common/functions";
+import { SetType } from "../../../../app/slices/main/types";
 import { ListItem, ListItemText, ListItemPrimaryText, ListItemSecondaryText, ListItemGraphic } from "@rmwc/list";
 import { IconButton } from "@rmwc/icon-button";
 import { Tooltip } from "@rmwc/tooltip";
