+++ src/components/main/views/list/ElementList.scss
