+++ src/components/main/views/list/ViewList.scss
